+++ app/src/main/kotlin/org/fossify/gallery/extensions/Activity.kt
@@ -79,7 +79,7 @@ fun Activity.openPath(path: String, forceChooser: Boolean, extras: HashMap<Strin
     openPathIntent(path, forceChooser, BuildConfig.APPLICATION_ID, extras = extras)
 }
 
-fun Activity.launchVideoPlayer(path: String, extras: HashMap<String, Boolean> = HashMap()) {
+fun Activity.launchGesturePlayer(path: String, extras: HashMap<String, Boolean> = HashMap()) {
     ensureBackgroundThread {
         val newUri = getFinalUriFromPath(path, BuildConfig.APPLICATION_ID)
         if (newUri == null) {